requirements.txt
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
crcmod==1.7
cryptography==36.0.1
docutils==0.18.1
idna==3.3
jmespath==0.10.0
lockfile==0.12.2
mysql-connector==2.2.9
oss2==2.15.0
psutil==5.9.0
pycparser==2.21
pycryptodome==3.12.0
pyDes==2.0.1
PyMySQL==1.0.2
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.8
